+++ src/test/java/entity/RatingTest.java
@@ -15,12 +15,12 @@ public class RatingTest {
     @Before
     public void init() {
         sampleStudent = new StudentUser("Sam", "4000");
-        rating = new Rating(sampleStudent, 10.0F);
+        rating = new Rating(sampleStudent, 10.0);
     }
 
     @Test
     public void getScoreOk() throws CommandNotAuthorizedException {
-        float desiredScore = 10.0F;
+        double desiredScore = 10.0;
         assertEquals(desiredScore, rating.getScore(), 0.001);
     }
 
@@ -28,7 +28,7 @@ public class RatingTest {
     public void getPoST() {
         StudentUser a = new StudentUser("Kevin", "kev123");
         a.setProgramDetail("DATA SCIENCE");
-        Rating r = new Rating(a, 10.0F);
+        Rating r = new Rating(a, 10.0);
         assertEquals("DATA SCIENCE", r.getRaterProgramOfStudy());
     }
